recode manual pages to UTF-8
I uploaded man-db 2.5.1-1 to incoming this morning, so I think we can
now proceed with the plan we discussed a while back:

  http://lists.debian.org/debian-i18n/2007/10/msg00063.html

I've attached a patch against git HEAD that recodes all manual pages to
UTF-8 on installation. I've tested this with debconf and it seems to be
doing the right thing. I suspect a small number of packages that have
inconsistently-encoded manual pages will fail to build as a result of
this, but I believe that these are rather few and I contend that they
were buggy anyway (as man would fail to display those pages).

The reason for the slightly baroque installation method (install file,
man --recode to .new, mv back) is that this allows man to guess the
encoding based on the installation directory.

Note that the original (non-.UTF-8) installation directories are
retained, per the latest discussion in #440420.