dh: Support debian/rules calling make with -B
That is useful to avoid issues with phony implicit rules
(see bug #509756).

Apparently make treats the name of the Makfile as an automaticall
set up target, so this causes it to try to build the Makefile
even though it's up-to-date, and the implicit target
makes it run 'dh debian/rules'.

So, make that a no-op.