remove _mkdir, use mkdir -p
_mkdir is not necessary, because mkdir's error messages are good enough
if a file by the name of the directory exists ("cannot create directory: File
exists"), or if a file is where the parent directory should be ("cannot
create directory: Not a directory")

Using mkdir -p seems useful, in case someone wants a deeply nested
builddir.

This also changes the return value of mkdir_builddir, but nothing currently
tests it.