remove verbose_print of buildsystem selection details
This would be the only place in debhelper where -v enables debugging info
that is not just shell commands being run.

Since --list can be used to see details of build system selection, and
since it will probably be obvious which one is selected in -v mode due to
the commands run, I think this oddity is unnecessary.