Ensure make doesn't print directories when checking for target existance.
Whenever make is run with --print-directory option, make -C sometimes print
Entering/Leaving directory messages to stdout even with -s in effects  This
breakes a check for target existance as it relies on make printing nothing when
target does not do anything. Hence explicitly pass --no-print-directory to make
to avoid it.

Signed-off-by: Modestas Vainius <modestas@vainius.eu>