Generate and install dh_auto docs when building debhelper.
Since dh_auto_* PODs and dh_auto.pod contain placeholders and final versions
are generated with the use of podselect, they are not directly translatable.
However, the final postprocessed PODs (with man/dh_auto_pod) are translatable.
So these temporary PODs are written to the man/ subdir and they could be added
to po4a as needed. However, I do not add those PODs for translation in this
commit (neither previous versions of dh_auto_* were there).