drop regex support from remove_command_options
I hope that it will not be needed; indeed I doubt that
remove_command_options will be used much, because sequence addons would
need to try to do conflicting things to need it. And the interface makes it
hard for such conflicting sequence addons to work around the other, since
addons can be loaded in either order. So let's not encourage them too
much, and if there's a use case later, we can made changes.

I haven't applied Modestas's enhanced patch that allows adding an option to
all commands because I similarly think it might not be used. If a use case
comes along we can add something like that.