Enable verbose ctest output on test failure
When test fails, enable verbose ctest output. This allows to
get more details on a test failure from the build logs.

Auto-select cmake in further steps only if cmake was run in configure step.
CMake writes CMakeCache.txt to build directory when it is run. Depend on the
presence of this file for auto-selection in build, test, install and clean
steps.

Signed-off-by: Modestas Vainius <modestas@vainius.eu>