makefile buildsystem: Chomp output during test for full compatability with debhelper 7.4.11. Closes: #570503
Previously the test used make -s -n | head -n 1 and then chomped the
output. In the case of this bug, root-system's Makefile *always* outputs
something to stdout, even for targets that don't exist, before configure is
run. It accidentially worked before, since the first line it outputs
happens to be empty.

So bring back the chomp to retain compatability with this package that used
to work before, but the test only does the right thing for this package due
to sheer luck, really.