dh_testversion: Removed this deprecated command.
2 packages in the archive still use it, but both like this:

ifneq ($(shell dh_testversion '<<' 7.4.10 2>/dev/null && echo old),old)
dh_parallel = --parallel
endif

Happily, the way that was written, if the command doesn't exist,
it does the right thing and enables the behavior for the new debhelper.
Not that it would greatly matter if it did not, since --parallel is not
crucial.