dh_strip: make stripping static libraries deterministic
Previously, calling strip on a static library would reset the timestamp
of every file in the archive to the current time, and reset the permissions
of every file based on the umask.

Now, the --enable-deterministic-archives is passed to strip to cause strip
to write the .a file with zero timestamps and consistent file permissions.