dh_builddeb: Use explicit -Z arg for ddebs to avoid FTBFS
When maintainer passes (e.g.) -Zgzip to dpkg-deb via dh_builddeb, then
we cannot pass that AND -Sextreme to dpkg-deb as gzip does not support
the "extreme" compression strategy.  For now, simply use the
compression parameters given by the maintainer for the .ddeb as well.

It is not clear to me that this is truly the best long term solution.
The maintainer probably only wanted the -Z to apply for the actual deb
and not the ddeb.

Reported-by: Reiner Herrmann <reiner@reiner-h.de>
Signed-off-by: Niels Thykier <niels@thykier.net>