Define SOURCE_DATE_EPOCH when setting build flags
When SOURCE_DATE_EPOCH is defined software which needs a timestamp can use it
as a reference date instead of the current date and time. For more details,
see: https://wiki.debian.org/ReproducibleBuilds/TimestampsProposal

The variable is set alongside other build flags, and its value is the time
of the latest entry in `debian/changelog`.

Closes: #791823