dh_auto_test: Add VERBOSE=1 when using autoconf
When the build system is autoconf, assume that one also uses automake,
and pass VERBOSE=1 to make check too. Without VERBOSE=1, newer
automake-generated makefiles will not display the actual test errors,
but store them in a file. This makes build logs considerably less useful
when it comes to discovering test failures.

With VERBOSE=1 set, test failures are displayed. It should have no ill
effects when used with a non-automake makefile + autoconf combination.

The test suite was updated to reflect the changes, and new tests were
added to verify the new functionality.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>
Signed-off-by: Niels Thykier <niels@thykier.net>