dh_clean: Work around for cdbs d/compat auto-vivification
cdbs has a feature where it will create debian/compat and set its
contents to 5 if it does not exist (and DH_COMPAT does not exist).

The issue being builds generally call "clean" before "build" and as
such packages relying on this feature now FTBFS.  As they were not
listed as packages in danger (lintian considered them to have compat 5
or greater), we now instate this work around to cater for them.

Signed-off-by: Niels Thykier <niels@thykier.net>