dh_installmenu: Skip maintscripts for non-exec menu files
Regular non-executable files are handled just fine by the menu
trigger, so stop the creating maintainer scripts for them.

Bonus info: There is currently only 1 executable menu file in the
archive and it does not need the maintscript either.  Nor does it have
one (because it does not use debhelper to install it).

Signed-off-by: Niels Thykier <niels@thykier.net>