dh_strip: disable file(1) cache when not using build-ids
It breaks in fun and exciting ways for hardlinked ELF binaries if we
are not using build-ids.  For now just disable the file(1) caching
in that case as it is not critical and an uncommon case.

The build-id case works (thanks to Sven Joachim) since the deattached
debug symbol is in a predictable location we can test for.

Signed-off-by: Niels Thykier <niels@thykier.net>