Remove myself from uploaders
Niels is doing a really, really great job.  I originally wanted to just
be sure there was _someone_ on the package, even if it was just me in
way over my head, and IMO we got waaay more than that in Niels. :)

I'm planning to stay on the mailing list, but having me in Uploaders
really doesn't reflect any form of reality. :D