fix autoconf/cross regression from #836988
When adding the makefile buildsystem cross variables, the intention was
that it would not affect non-makefile buildsystems (in particular no
downstream buildsystems). However, the decision which buildsystem to use
is done on a per-target basis. Thus a typical autoconf package will use
the autoconf buildsystem for configure and test, but fall back to the
makefile buildsystem for clean and build. Thus the cross variables were
added for autoconf build as well, which broke the cross build of icu.

The solution chosen here is to have autoconf take over build and clean
from makefile as well by inheriting its methods. Thus the semantics stay
unchanged with the exception of not adding the cross variables for
build.

All other children of the makefile buildsystem (including mkcmake)
already take over all targets, so this issue really only affects
autoconf.

Fixes: 7ea67c9aace4 ("makefile.pm: Set CC+CXX to the host compilers when
       cross-building")
Signed-off-by: Helmut Grohne <helmut@subdivi.de>
Signed-off-by: Niels Thykier <niels@thykier.net>