pkgfile: Attempt to make it scale better for more pkgs
Each time pkgfile is called, it will glob the debian/ dir (with a
filter).  Obviously this adds up, but unfortunately, glob is a lot
cheaper in some cases than calling buildarch()/buildos(), which
requires Dpkg::Arch.

To solve this, we implement a cache of the globbing so we at most glob
once per file name.  Depending on that result, we will now
unconditionally check for architecture qualified files or not.

This should make both cases faster for 2+ packages.  It should also
apply to dh when checking if it can skip a command.  When testing
about 500 empty "transitional" packages, we save about 7-8% run time
for dh_install.

Signed-off-by: Niels Thykier <niels@thykier.net>