Dh_Lib: Make -i/-a to avoid Getopt::Long
If we need to use Getopt::Long, we pay a ~0.020s overhead (per
helper).  When doing dpkg-buildpackage -A/-B builds, dh will
unconditionally pass -i/-a in those cases making each helper cost a
bit more.

With this change, we short-cut -i and -a (but not their long variants)
so Dh_Lib can handle those directly.  This reduces the overhead for
-A/-B so they are comparable to -b.

Signed-off-by: Niels Thykier <niels@thykier.net>