pkgfile: Create a "secret" bulk check call for dh
With "enough" packages, pkgfile is still a very expensive part of dh.
Mostly due to the can_skip check which often need to call pkgfile to
determine if a helper can be skipped.  By doing a "bulk check" we can
reduce that overhead by about roughly 50% (20-25% of dh's total
runtime).

Signed-off-by: Niels Thykier <niels@thykier.net>