dh_fixperms: Add @mode_0755_patterns (for derivatives)
Debian has nothing general patterns that must be chmod'ed to 0755, but
DilOS needs "*.so" and "*.so.*" to be 0755 (instead of 0644).  As it
only takes 2-3 lines of code to mode the patterns out of the logic,
then lets have an empty list, so DilOS's changes can be smaller and
limited to some data lists.

Signed-off-by: Niels Thykier <niels@thykier.net>