Re-instate bug compatibility to fix #867866
Previous versions of debhelper had "interesting" way of interpreting
arguments where it split them on space.  This was mistakenly
"corrected" in 10.6 when debhelper started to use bsd_glob (that does
not split on spaces).

To resolve this, manully split arguments like the original code would
(effectively) have done so packages are unaffected by it.

Signed-off-by: Niels Thykier <niels@thykier.net>