Dh_Getopt: Fix discard of warning/error messages in option parsing
When testing -O options, the __WARN__ handler would be set to a sub to
throw away warnings for -O options.  However, the original __WARN__
handler would not be properly reset in the case where it was
originally "undef" (which happens to be the default).  This triggered
a discard of real error messages in case there was a later call to
"getoptions".

Concretely, this was trivially reproducible via:

  DH_OPTIONS=-Oa dh_makeshlibs -ppackage-that-does-not-exist

This /should/ have failed with:

  dh_makeshlibs: Requested unknown package foo via -p/--package, expected one of: [...]
  dh_makeshlibs: unknown option or error during option parsing; aborting

But prior to this commit, only the latter line would appear.  This
sadly makes the error rather hard to understand.

Signed-off-by: Niels Thykier <niels@thykier.net>