Add DH_EXTRA_ADDONS env variable to specify local addons.
This is intended to be used by downstreams or specific local configurations
that require a debhelper addon to be run during multiple builds without
having to patch a large number of rules file. If at all possible, this should
be avoided in favor of a --with flag in the rules file.

Signed-off-by: Niels Thykier <niels@thykier.net>