dh_strip: Fix bug generating Build-Id field with --dbg-package
The problem occurred because we overrode the list of recorded
build-ids after each package.  This works perfectly fine for dbgsym
packages as they always have a 1:1 with the original package.
However, a manual --dbg-package can host debug symbols for multiple
binary packages.

There are two paths to triggering this issue:

  dh_strip -p pkg1 -p pkg2 --dbg-package=foo-dbg

and:

  dh_strip -p pkg1 --dbg-package=foo-dbg
  dh_strip -p pkg2 --dbg-package=foo-dbg

  (This case was observed in krb5/1.15.2-2 where it loops over most of
  the packages and calls dh_strip on each of them with the same
  --dbg-package parameter.)

They fundamentally have the same issue and fix but the second case is
slightly more involved (as we have to merge with the previous run).

Signed-off-by: Niels Thykier <niels@thykier.net>