cmake.pm: Place path-to-source at end of command line invocation.
This allows params in debian/rules of the form "-D FOO=BAR" instead of
only "-DFOO=BAR" (without quotes).  cmake permits both forms, but
mishandles the directory if any of the params contain a space between
the -D and the var/value.  If the directory is last, it always works.