dh_installsystemd: Simplify error handling
The intention was to carry on all possible work and report missing
unit files specified on the command line at the end of execution.
However, now the scripts fails immediately when it tries to read the
content of the unit filed and the late check is never executed. Remove
the late check.