dh_installinit: --name implies, that init script is present
Previously, `dh_installinit' silently did nothing, when --name option
was passed, but initscript debian/<package>.<name>.init was not found.

In almost all cases, explicit --name means that package maintainer meant
to install init script. If it is not present, it is bug, and must not be
hidden. Now, error is reported in this case. (Closes: #462389)

Signed-off-by: Niels Thykier <niels@thykier.net>