dh_builddeb: Remove support for --destdir
It is not useful as the packager cannot set it to anything as it is
the builder that has an expectation of where the debs are produced.
It would have been a lot more useful if the builder could have
specified this value some how (but it would have to be aligned with
dpkg with e.g.  for the .changes).

Signed-off-by: Niels Thykier <niels@thykier.net>