dh: Skip some invalid calls to helpers after an override
With the new "conditional" addon feature, it was possible to trigger a
gratious useless warning with a conditional override.  Consider:

    override_dh_foo-arch:
        ...

And dh_foo was added by an arch-only addon.  Before this commit, we
would do:

    debian/rules override_dh_foo-arch
    dh_foo -a -N<for each arch:any package>

The latter call to dh_foo rightfully complained there was nothing to
do.  With this commit, we correctly omit the latter call.

Signed-off-by: Niels Thykier <niels@thykier.net>