dh: Fix regression in dh addon API
Previously, dh's API functions were sloppy in their return values but
*often* they would return a truth value.  We changed that in 12.5 by
insisting on return undef.  However, that broke addons relying on the
API function to work around Perl's idiom of "module must end with a
truth value".

To resolve that, we now always return a truth value from all API
functions to fix that.  This also give consistent behaviour, which is
better than the previous setup.

Signed-off-by: Niels Thykier <niels@thykier.net>