autoconf.pm: Do not check for unknown --disable/enable/with/without-* options
By default autoconf will check for unknown --disable/enable/with/withouth-*
options and will emit a warning on standard error. This can be tuned by
the upstream maintainer with the AC_DISABLE_OPTION_CHECKING macro, to
make these fatal, or to not perform the checks by default.

Given that we unconditionally pass possibly unknown options to the
configure scripts, we should disable these checks to avoid emitting
warnings, which are at least confusing, and at worse might trip over
CI and test suite that check for output on standard error as markers
for failure.

Signed-off-by: Guillem Jover <guillem@debian.org>
Signed-off-by: Niels Thykier <niels@thykier.net>