Optimize imports for addon sequences
In a future commit, we will move the addon API to a separate module
and that will imply that missing imports in addons will become
visible.  This commit fixes the built-in addons to work in the new
conditions (most of them already did) and replaces "wildcard imports"
with importing exactly what we need.

Signed-off-by: Niels Thykier <niels@thykier.net>