dh_missing: Check for possible alternative install paths
Detect missing files being installed from different paths and warning
specifically about this.  This should hopefully reduce the number of
bug reports where people do not understand by dh_missing reports
missing files when they feel it is not missing.

The listed bug is not the first of its kind and is unlikely to be the
last without this change.

Signed-off-by: Niels Thykier <niels@thykier.net>