dh_clean: Avoid corner-case in restore_all_files
If dh_clean is aborted during clean, we might trigger a race where
some of the files from the restore bucket has been deleted but the
index has not.  This creates an error where dh_clean will fail as it
cannot restore a file that it already restored and deleted itself.

We fix this by deleting the index immediately after restore to ensure
that a rerun of dh_clean will skip the restore.

Signed-off-by: Niels Thykier <niels@thykier.net>