systemd: Only merge /lib/systemd/system into /usr for now
systemd and init-system-helpers read both /lib/systemd/system and
/usr/lib/systemd/system for units, so it's OK to normalize everything
to be installed into /usr/lib/systemd/system.

However, the same is not true for other subdirectories of /lib/systemd,
such as /lib/systemd/system-generators. Until the systemd maintainers
confirm which files they would like to be migrated, and whether a
${misc:Depends} on a newer systemd is required for the transition, it
seems like the safest thing to do is to restrict this merging to just
the units from /lib/systemd/system.

Closes: #992554
Signed-off-by: Simon McVittie <smcv@debian.org>