Stop services on upgrade during preinst with --no-restart-after-upgrade
There are two bugs addressed in this commit: #989155 which involves
packages that move from --restart-after-upgrade to
--no-restart-after-upgrade, and #994204 which involves packages that
rely upon --no-stop-on-upgrade being restarted regardless.

I believe the root of the issue lies in #989155 and the manner in which
it was fixed. Specifically, the root of the problem in #989155 is that,
when "--no-restart-after-upgrade" is requested (indicating that the
restart of services should happen *during* the upgrade, stopping before
files are replaced and starting again afterwards, rather than the
default of *after* the upgrade), the responsibility for stopping and
starting the services is split across *two* versions of the package.

The stopping of services is handled by the prerm of the *prior* version
of the package, while the (re)starting of services is handled by the
postinst of the *new* version of the package. This breaks the ability to
move from one restart behaviour (--restart-after-upgrade) to another
(--no-restart-after-ugprade). The solution presented here is to instead
move the stopping of services to the preinst script of the *new* version
of the package in the case that --no-restart-after-upgrade is requested.