docs: Replace "DEB_MAINT_BUILD_OPTIONS" with "DEB_BUILD_MAINT_OPTIONS"
debhelper's documentation was referring to an environment variable
"DEB_MAINT_BUILD_OPTIONS", which is used by dpkg-buildflags(1). However,
dpkg-buildflags(1) is using the environment variable
"DEB_BUILD_MAINT_OPTIONS" (mind that the second and third part of the
variable's name are exchanged). This commit fixes debhelper's
documentation by replacing all occurrences of the old name with the new
name.