Move compat upgrade checklist to its own file
Make it easier for people to find the compat upgrade checklist by
moving its to own file, so they do not have to "infinity" scroll
through debhelper(7) to find it.

To aid people working from memory, the old section header is left in
place with a link to the new document.

Signed-off-by: Niels Thykier <niels@thykier.net>