Dh_Lib.pm: Support `?=` in substvar files
Hopefully, a future version of dpkg will support this syntax to mark a
substvar as optional. That would allow helper tools to provide
substvars more liberally as they would not trigger warnings when
unused.

This commit is a step towards supporting the new syntax by recognizing
it as valid and preserving the original assignment operator when
someone else created the substvar.

Signed-off-by: Niels Thykier <niels@thykier.net>