Dh_Lib.pm: _strip_spaces now explicitly returns undef on undef argument.
_strip_spaces sometimes get's used to process optional source fields
such as for example the 'Section:' field. In this case, the argument
to _strip_spaces would be undef and without this patch perl issues a
warning message.

This patch makes _strip_space simply return undef in this case, assuming
that the caller of _strip_spaces is prepared to handle missing (optional)
source field as undef.

See bugs.debian.org: #1010594

Signed-off-by: Niels Thykier <niels@thykier.net>