dh_installman: fix regex to parse section out of filename
If a filename has multiple dot-digit sequences, the regex fails to
extract the last one, which is the section.

For example the section of libmanpage.so.9.2.3 is extracted as 9.2.3.

Fix the regex to match on the last dot and single digit
in the file name and letters following it.
This ensures we don't glob up 1.2.3, and also support .3perl and such.