docs: convert PROGRAMMING to Markdown
I find reading plain text documents a bit hard, especially on smaller
screens, and converting the file to Markdown allows easy conversion to
HTML, while not reducing plain text readability.

This Markdown file is mostly compatible with the [original syntax][1],
except for the table and the links (even though the [last beta][2] of
the original markdown.pl does support the latter).

I don't particularly like the table in the "Arguments processing"
section, as it requires GFM / Markdown Extra extensions; it also doesn't
look great, as you can't inset [newlines in tables][3].

The simplest way to convert this file from Markdown to HTML locally
is the following:

    # apt install cmark-gfm
    $ cmark-gfm --extension table PROGRAMMING.md > PROGRAMMING.html

[1]: https://daringfireball.net/projects/markdown/syntax
[2]: https://sources.debian.org/src/markdown/1.0.2~b8-7/ChangeLog/#L83
[3]: https://github.com/markdown-it/markdown-it/issues/406