Reject debhelper config files without a package prefix for mulit binary packages in compat 15
Starting in compat 14, it will now trigger a warning to have
`debian/install` instead of `debian/package.install` when
`debian/control` contains 2 or more binary packages. In compat 15,
this is promoted to an error.  The main exceptions are files like
`debian/changelog`, `debian/NEWS`, and `debian/copyright`, which are
applied to all packages and can still be used without a prefix.

Closes: #1014537
Gbp-Dch: Full
Signed-off-by: Niels Thykier <niels@thykier.net>