Resolve divergence of vcs-git and dgit
The debhelper source for 13.11.5 as uploaded to Debian differs from what
is recorded in this git tree. What has been uploaded matches the dgit
tree. In the upload, a few translation files are reformatted/regenerated
to match source changes. This commit updates them in the git history to
match what actually was uploaded.

Gbp-Dch: ignore
Signed-off-by: Helmut Grohne <helmut@subdivi.de>