Replace `d/compat` with a `X-DH-Compat` field
Starting debhelper/14, `debian/compat` can no longer be used to
specify the compat level. To avoid breakage of packages that have
alredy started using compat 14, the error triggers when
debhelper-compat 14 becomes stable.

With this, the debhelper compat level will always be in the
debian/control file, once debhelper/14 is released.

Signed-off-by: Niels Thykier <niels@thykier.net>