Make it easier to re-bootstrap without `dwz`
The t64 transition saw `dwz` become uninstallable on a 32-bit
architecture, which in turn made `debhelper` uninstallable. This would
have required a re-bootstrap via cross-building, or a "dirty"
debhelper without `dwz` support.

With this commit, we make the "dirty" debhelper case easier by
explicitly providing Build-Profiles for this case. Consumers should be
aware that the "dirty" debhelper will not produce bit-for-bit
reproducible results with the "full" debhelper.

The commit adds a similar feature for `man-db`. For this part,
`debhelper` already self-bootstraps without that tooling, so this was
just a matter of a build-profile.

Signed-off-by: Niels Thykier <niels@thykier.net>